% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_tibble.R
\name{model_tibble}
\alias{model_tibble}
\title{Wrapper for model.matrix}
\usage{
model_tibble(formula, ...)
}
\arguments{
\item{formula}{An object of class "formula". Describes the model
specification}

\item{...}{Other arguments corresponding to the parameters of model.matrix}
}
\value{
Tibble of model.matrix
}
\description{
Wrapper for model.matrix
}
