% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ridge.R
\name{ridge_estimator}
\alias{ridge_estimator}
\title{Estimates the ridge model}
\usage{
ridge_estimator(formula = NULL, lambda = 0, data = NULL)
}
\arguments{
\item{formula}{An object of class "formula". Describes the model
specification}

\item{lambda}{A number which represents the penalty. Default is set to zero
which yields the least squares estimator}

\item{data}{An optional data frame or tibble which contains the corresponding
data. When not specified the data will be taken from the parent environment
of "formula".}
}
\description{
Estimates the ridge model
}
