% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RidgeLassoTools-package.R
\docType{package}
\name{RidgeLassoTools-package}
\alias{RidgeLassoTools}
\alias{RidgeLassoTools-package}
\title{RidgeLassoTools: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: Jonas Veit \email{first.last@example.com} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

}
\keyword{internal}
